refactor(audio_tools): Introduce config struct for gen_filterbank
This commit refactors the `gen_filterbank` function to accept a `filterbank_config_t` struct. This improves the function's flexibility and makes it more extensible.

To provide a convenient way to create a default configuration, a preprocessor macro was initially considered. However, this approach led to compilation issues with complex arguments. I ended up choosing to implement a `static inline` function, `get_default_filterbank_config`, to avoid this issues.

This refactoring includes:
- Addition of `filterbank_config_t` and related enums.
- Implementation of the `get_default_filterbank_config` inline function.
- Updates to `main.c` and `tests/test_onset.c` to use the new configuration struct.
- Minor warning fixes for unused variables and format specifiers to improve overall code quality.